refactor: 질문 꾸러미 상세 페이지 스타일 파일 분리
