+++ src/pages/interview/InterviewReady/style.ts
